backend/requirements.txt
aiofiles==24.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
Faker==37.5.3
fastapi==0.116.1
fastapi-limiter==0.1.6
GeoAlchemy2==0.18.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
